chore(api): genergate api document buildtime
