+++ apps/api/src/config/document.config.ts
@@ -1,0 +1,20 @@
+import type { INestApplication } from "@nestjs/common";
+import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger";
+import type { OpenAPIObject } from "@nestjs/swagger";
+
+export async function setupSwagger(app: INestApplication): Promise<OpenAPIObject> {
+  const config = new DocumentBuilder()
+    .setTitle("time-align API")
+    .setDescription("time-align API")
+    .setVersion("1.0")
+    .addBearerAuth()
+    .build();
+
+  const document = SwaggerModule.createDocument(app, config);
+
+  SwaggerModule.setup("api-docs", app, document, {
+    jsonDocumentUrl: "/api-json",
+  });
+
+  return document;
+}
