+++ apps/api/src/main.ts
@@ -1,6 +1,6 @@
 import { NestFactory } from "@nestjs/core";
-import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger";
 import { AppModule } from "./app.module";
+import { setupSwagger } from "./config/document.config";
 
 // https://docs.nestjs.com/recipes/hot-reload#hot-module-replacement
 // biome-ignore lint/suspicious/noExplicitAny: <explanation>
@@ -9,17 +9,7 @@ declare const module: any;
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
-  const config = new DocumentBuilder()
-    .setTitle("time-align API")
-    .setDescription("time-align API")
-    .setVersion("1.0")
-    .build();
-
-  const document = SwaggerModule.createDocument(app, config);
-
-  SwaggerModule.setup("api-docs", app, document, {
-    jsonDocumentUrl: "/api-json",
-  });
+  await setupSwagger(app);
 
   await app.listen(process.env.PORT ?? 3001);
